# docs/auth_service_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Saturn MouseHunter Auth Service",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/menus/tree": {
      "get": {
        "tags": [
          "菜单权限"
        ],
        "summary": "Get Menu Tree",
        "description": "获取完整菜单树结构\n\n**权限要求**: 已认证用户\n\n**用途**:\n- 管理界面显示所有菜单\n- 权限配置界面",
        "operationId": "get_menu_tree_api_v1_menus_tree_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MenuTree"
                  },
                  "type": "array",
                  "title": "Response Get Menu Tree Api V1 Menus Tree Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/users/{user_id}/menus": {
      "get": {
        "tags": [
          "菜单权限"
        ],
        "summary": "Get User Menus By Id",
        "description": "获取指定用户的可访问菜单\n\n**权限要求**: 管理员用户或查看自己的菜单\n\n**参数**:\n- user_id: 目标用户ID",
        "operationId": "get_user_menus_by_id_api_v1_users__user_id__menus_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMenuResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MenuTree": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "permission": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permission"
          },
          "menu_type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MenuType"
              }
            ],
            "default": "menu"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 0
          },
          "is_hidden": {
            "type": "boolean",
            "title": "Is Hidden",
            "default": false
          },
          "meta": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/MenuTree"
            },
            "type": "array",
            "title": "Children",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "title"
        ],
        "title": "MenuTree",
        "description": "菜单树结构"
      },
      "MenuType": {
        "type": "string",
        "enum": [
          "menu",
          "button",
          "tab"
        ],
        "title": "MenuType",
        "description": "菜单类型"
      },
...
      "UserMenuResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "用户ID"
          },
          "user_type": {
            "type": "string",
            "title": "User Type",
            "description": "用户类型"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions",
            "description": "用户权限列表"
          },
          "menus": {
            "items": {
              "$ref": "#/components/schemas/MenuTree"
            },
            "type": "array",
            "title": "Menus",
            "description": "可访问菜单"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新时间"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "user_type",
          "permissions",
          "menus",
          "updated_at"
        ],
        "title": "UserMenuResponse",
        "description": "用户菜单响应模型"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}